requirements.txt
contentful
contentful-management
python-dotenv
python-dateutil
ipython
